Fix promise and arrow function usage in intro example.
`serial.open().then()` is not a good idea. The `.then()` does nothing, and any errors that occur are ignored. Instead, the promise returned by `serial.open()` should be returned inside the fulfillment handler of the `SerialPort.requestList()` promise. You can then add a failure handler with `.catch`, which will handle any errors that occur in either: requesting the list, processing the list, or opening the serial port.

Also, ES6 arrow functions do not need parenthesis in their single-argument form, and you can leave out braces and return if you just want to return an expression (as in the `filter` case).