Introduce Navigator and WorkerNavigator partial interfaces
The interfaces provide a `serial` attribute that is the entrypoint to
the API. The getPorts() and requestPort() methods replace the SerialPort
constructor in order to avoid exposing the raw serial port path to
script. The Serial interface also provides events that can be fired when
a port is added or removed from the system as will be the case for USB
based serial devices.